Create README.md
I'm a front end developer in Ethiopia. I'm mostly focus on the functionality of the website. currently i'm an IT student in Wolllo University @ Kombolcha Institute of Technology, which is one of the few technology campuses in Ethiopia. if you anything challenging that could help me grow as a website developer i'm happy to be challenged.